Refactor decoder to read from an interface instead of a string
Using a string as input to the decoder meant it was required to read the whole length of data to
decode into memory, which meant decoding was a potentially very heavy on memory. This change
abstracts out the input to the decoder, and adds a DataSource called String to replace it. This
change breaks backwards compatibility.